# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: resume
    description: Resume analysis operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /resume/cover-letter:
    post:
      operationId: generateCoverLetter
      tags: [resume]
      summary: Generate a tailored cover letter
      description: Uses AI to generate a personalized cover letter based on the resume and job description
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeResumeBody"
      responses:
        "200":
          description: Cover letter generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoverLetterResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /resume/career-roadmap:
    post:
      operationId: generateCareerRoadmap
      tags: [resume]
      summary: Generate a career roadmap
      description: Uses AI to create a phased career roadmap based on skill gaps and the target role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CareerRoadmapBody"
      responses:
        "200":
          description: Career roadmap generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CareerRoadmapResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /resume/analyze:
    post:
      operationId: analyzeResume
      tags: [resume]
      summary: Analyze resume against job description
      description: Uses AI to analyze a resume against a job description and returns a structured report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeResumeBody"
      responses:
        "200":
          description: Analysis complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    AnalyzeResumeBody:
      type: object
      properties:
        resumeText:
          type: string
          description: Extracted text from the uploaded resume
        jobDescription:
          type: string
          description: The full job description text
      required:
        - resumeText
        - jobDescription
    WeakBullet:
      type: object
      properties:
        original:
          type: string
        improved:
          type: string
      required:
        - original
        - improved
    RedFlag:
      type: object
      properties:
        issue:
          type: string
        fix:
          type: string
      required:
        - issue
        - fix
    AnalysisSummary:
      type: object
      properties:
        overallImpression:
          type: string
        keyStrengths:
          type: string
        priorityImprovements:
          type: string
        finalRecommendation:
          type: string
        confidenceLevel:
          type: string
          enum: [High, Medium, Low]
      required:
        - overallImpression
        - keyStrengths
        - priorityImprovements
        - finalRecommendation
        - confidenceLevel
    AnalysisResult:
      type: object
      properties:
        matchScore:
          type: number
          minimum: 0
          maximum: 100
        matchedKeywords:
          type: array
          items:
            type: string
        missingKeywords:
          type: array
          items:
            type: string
        weakBullets:
          type: array
          items:
            $ref: "#/components/schemas/WeakBullet"
        redFlags:
          type: array
          items:
            $ref: "#/components/schemas/RedFlag"
        summary:
          $ref: "#/components/schemas/AnalysisSummary"
      required:
        - matchScore
        - matchedKeywords
        - missingKeywords
        - weakBullets
        - redFlags
        - summary
    CoverLetterResult:
      type: object
      properties:
        coverLetter:
          type: string
          description: The generated cover letter text
        wordCount:
          type: number
      required:
        - coverLetter
        - wordCount
    CareerRoadmapBody:
      type: object
      properties:
        resumeText:
          type: string
        jobDescription:
          type: string
        missingKeywords:
          type: array
          items:
            type: string
        matchScore:
          type: number
      required:
        - resumeText
        - jobDescription
        - missingKeywords
        - matchScore
    RoadmapPhase:
      type: object
      properties:
        phase:
          type: string
        duration:
          type: string
        goal:
          type: string
        tasks:
          type: array
          items:
            type: string
        resources:
          type: array
          items:
            type: string
      required:
        - phase
        - duration
        - goal
        - tasks
        - resources
    CareerRoadmapResult:
      type: object
      properties:
        targetRole:
          type: string
        summary:
          type: string
        phases:
          type: array
          items:
            $ref: "#/components/schemas/RoadmapPhase"
        estimatedTimeToReady:
          type: string
      required:
        - targetRole
        - summary
        - phases
        - estimatedTimeToReady
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message