# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Uses AI to generate a personalized cover letter based on the resume and job description
 * @summary Generate a tailored cover letter
 */
export const GenerateCoverLetterBody = zod.object({
  resumeText: zod.string().describe("Extracted text from the uploaded resume"),
  jobDescription: zod.string().describe("The full job description text"),
});

export const GenerateCoverLetterResponse = zod.object({
  coverLetter: zod.string().describe("The generated cover letter text"),
  wordCount: zod.number(),
});

/**
 * Uses AI to create a phased career roadmap based on skill gaps and the target role
 * @summary Generate a career roadmap
 */
export const GenerateCareerRoadmapBody = zod.object({
  resumeText: zod.string(),
  jobDescription: zod.string(),
  missingKeywords: zod.array(zod.string()),
  matchScore: zod.number(),
});

export const GenerateCareerRoadmapResponse = zod.object({
  targetRole: zod.string(),
  summary: zod.string(),
  phases: zod.array(
    zod.object({
      phase: zod.string(),
      duration: zod.string(),
      goal: zod.string(),
      tasks: zod.array(zod.string()),
      resources: zod.array(zod.string()),
    }),
  ),
  estimatedTimeToReady: zod.string(),
});

/**
 * Uses AI to analyze a resume against a job description and returns a structured report
 * @summary Analyze resume against job description
 */
export const AnalyzeResumeBody = zod.object({
  resumeText: zod.string().describe("Extracted text from the uploaded resume"),
  jobDescription: zod.string().describe("The full job description text"),
});

export const analyzeResumeResponseMatchScoreMin = 0;
export const analyzeResumeResponseMatchScoreMax = 100;

export const AnalyzeResumeResponse = zod.object({
  matchScore: zod
    .number()
    .min(analyzeResumeResponseMatchScoreMin)
    .max(analyzeResumeResponseMatchScoreMax),
  matchedKeywords: zod.array(zod.string()),
  missingKeywords: zod.array(zod.string()),
  weakBullets: zod.array(
    zod.object({
      original: zod.string(),
      improved: zod.string(),
    }),
  ),
  redFlags: zod.array(
    zod.object({
      issue: zod.string(),
      fix: zod.string(),
    }),
  ),
  summary: zod.object({
    overallImpression: zod.string(),
    keyStrengths: zod.string(),
    priorityImprovements: zod.string(),
    finalRecommendation: zod.string(),
    confidenceLevel: zod.enum(["High", "Medium", "Low"]),
  }),
});