# lib/api-zod/src/generated/types/careerRoadmapResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RoadmapPhase } from "./roadmapPhase";

export interface CareerRoadmapResult {
  targetRole: string;
  summary: string;
  phases: RoadmapPhase[];
  estimatedTimeToReady: string;
}