# lib/api-zod/src/generated/types/coverLetterResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CoverLetterResult {
  /** The generated cover letter text */
  coverLetter: string;
  wordCount: number;
}